Use a different token for accessing packages that for accessing the repo
On repos with branch protection, the token used to access the repo
must be a PAT of a technical user who has the permission to
bypass the PR requirement. This cannot be done for the `github.token`.
However, the `github.token` is enough access the package registry.
Separating the two allows a less powerful token to be used as the PAT.