Use a PAT instead of the automatic token authentication
because the latter will not trigger subsequent workflow to
notify the dependant repos.

When you use the repository's GITHUB_TOKEN to perform tasks, events
triggered by the GITHUB_TOKEN, with the exception of workflow_dispatch
and repository_dispatch, will not create a new workflow run.
This prevents you from accidentally creating recursive workflow runs.
For example, if a workflow run pushes code using the
repository's GITHUB_TOKEN, a new workflow will not run even when the
repository contains a workflow configured to run when push events occur.

https://docs.github.com/en/actions/security-guides/automatic-token-authentication